parsers/hl7_parser.py
"""
HL7 v2 Parser — healthcare messaging standard.

THIS IS YOUR STUB TO COMPLETE.

HL7 v2 is the most widely used healthcare messaging standard — it handles
ADT (admit/discharge/transfer), ORU (lab results), ORM (orders), and more.

To complete this parser:
    pip install hl7apy

HL7 v2 message structure:
    MSH|^~\\&|SENDER|FACILITY|RECEIVER|DEST|20240101120000||ADT^A01|MSG001|P|2.5
    EVN|A01|20240101120000
    PID|1||MRN-001^^^HOSP^MR||SMITH^JOHN^A||19850315|M|||123 MAIN ST^^DALLAS^TX^75201
    PV1|1|I|ICU^101^A|||ATT-001^DOE^JANE|||SUR||||ADM|20240101

Common message types:
    ADT^A01  — patient admit
    ADT^A03  — patient discharge
    ADT^A08  — patient update
    ORU^R01  — lab / observation result
    ORM^O01  — order message
    ACK      — acknowledgement

Output: one row per message with key patient/visit/observation fields.

Usage (once implemented):
    parser = HL7Parser()
    df = parser.parse("path/to/messages.hl7")
    df = parser.parse(hl7_string)
"""

import logging
from typing import Any

import pandas as pd

from .base_parser import BaseParser, ParserError

logger = logging.getLogger(__name__)

# HL7 standard delimiters
FIELD_SEP    = "|"
COMPONENT_SEP = "^"
REPEAT_SEP   = "~"
ESCAPE_CHAR  = "\\"
SUBCOMP_SEP  = "&"


class HL7Parser(BaseParser):
    """
    HL7 v2.x message parser.

    Implement parse() below using hl7apy or hand-rolled segment parsing.
    The base_parser contract: take source, return pd.DataFrame.
    """

    @property
    def format_name(self) -> str:
        return "HL7 v2"

    def parse(self, source: Any) -> pd.DataFrame:
        """
        TODO: Implement this method.

        Suggested approach using hl7apy:

            from hl7apy.core import Message
            from hl7apy.parser import parse_message

            text = self._load(source)
            messages = self._split_messages(text)
            rows = []
            for raw_msg in messages:
                try:
                    msg  = parse_message(raw_msg, find_groups=False)
                    rows.append(self._extract_fields(msg))
                except Exception as e:
                    logger.warning(f"Skipping malformed HL7 message: {e}")

            return pd.DataFrame(rows)

        Key segments to extract:
            MSH  — message header (type, sender, timestamp, control ID)
            PID  — patient identification (MRN, name, DOB, gender, address)
            PV1  — patient visit (admit date, discharge date, location, doctor)
            OBX  — observation (lab results — value, units, reference range, status)
            OBR  — observation request (order ID, test code, specimen)
            DG1  — diagnosis (ICD code, description)
            IN1  — insurance information

        Suggested output columns:
            message_type, message_control_id, message_datetime,
            sending_facility, receiving_facility,
            patient_id, patient_name_last, patient_name_first,
            patient_dob, patient_gender, patient_address,
            admit_date, discharge_date, patient_class,
            attending_physician, patient_location,
            diagnosis_code, diagnosis_description,
            observation_id, observation_value, observation_units,
            observation_status, observation_datetime
        """
        raise NotImplementedError(
            "HL7Parser.parse() is not yet implemented. "
            "See the docstring above for implementation guidance. "
            "Install hl7apy with: pip install hl7apy"
        )

    def _load(self, source: Any) -> str:
        """Load raw HL7 text from file path, string, or bytes."""
        from pathlib import Path
        if isinstance(source, bytes):
            return source.decode("utf-8", errors="replace")
        if isinstance(source, Path):
            return source.read_text(encoding="utf-8")
        if isinstance(source, str):
            if len(source) < 512 and not source.strip().startswith("MSH"):
                path = Path(source)
                if path.exists():
                    return path.read_text(encoding="utf-8")
            return source
        raise ParserError(f"Unsupported source type: {type(source)}")

    def _split_messages(self, text: str) -> list[str]:
        """
        Split a file containing multiple HL7 messages.
        Each message starts with MSH.
        """
        messages = []
        current  = []
        for line in text.splitlines():
            line = line.strip()
            if not line:
                continue
            if line.startswith("MSH") and current:
                messages.append("\r".join(current))
                current = []
            current.append(line)
        if current:
            messages.append("\r".join(current))
        return messages

    def _safe_get(self, segment, field: int, component: int = 0, default: str = "") -> str:
        """
        Safely get a field value from a parsed HL7 segment.
        Implement this once you have hl7apy working.
        """
        try:
            val = segment.children[field]
            if component:
                return str(val.children[component].value)
            return str(val.value)
        except (IndexError, AttributeError):
            return default
