requirements.txt
# Core data
pandas>=2.1.0
pyarrow>=14.0.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0

# Database
psycopg2-binary>=2.9.9

# HTTP + networking
requests>=2.31.0
paramiko>=3.4.0     # SFTP

# Healthcare EDI (optional — needed for HL7)
# hl7apy>=1.3.4

# Avro (optional — needed for AvroParser)
# fastavro>=1.9.0

# Excel (optional — needed for CSVParser .xlsx support)
openpyxl>=3.1.2

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3     # retry decorator (used in outbound callers)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
